Update main.py
Add support for type bfloat16

+++ main.py
@@ -36,7 +36,7 @@ logger = HostLoggerAdapter(logging.getLogger(__name__), {})
 dict = {str: np.float32}
 print("Cтатус видеокарты: ", torch.cuda.is_available())
 device = "cuda:0" if torch.cuda.is_available() else "cpu"
-torch_dtype = torch.float16 if torch.cuda.is_available() else torch.float32
+torch_dtype = torch.bfloat16 if torch.cuda.is_available() else torch.float32
 
 model_id = "openai/whisper-large-v3"
